\name{odpc-package}
\alias{odpc-package}
\docType{package}

\title{One-Sided Dynamic Principal Components}

\description{
Computes One-Sided Dynamic Principal Components.
}
\details{
\tabular{ll}{
Package: \tab odpc\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2017-08-15\cr
Depends: \tab R (>= 3.3.1)\cr
License: \tab GPL (>= 2)\cr
Imports: \tab methods, Rcpp (>= 0.12.7), gdpc, forecast\cr
LinkingTo: \tab Rcpp, RcppArmadillo (>= 0.7.500.0.0)\cr
NeedsCompilation: \tab yes\cr
}

Index:
\preformatted{
components.odpcs    Get One-Sided Dynamic Principal Components from an odpcs object.
}
\preformatted{
odpc    Computes One-Sided Dynamic Principal Components.
}
\preformatted{
fitted.odpcs    Get reconstructed time series from an odpcs object.
}
\preformatted{
forecast.odpcs    Get forecasts from an odpcs object.
}
\preformatted{
plot.odpc   Plots an odpc object.
}
}
\author{
Daniel Peña, Ezequiel Smucler, Victor Yohai

Maintainer: Ezequiel Smucler <ezequiels.90@gmail.com>
}

\references{
Peña D., Smucler E. and Yohai V.J. (2017). “Forecasting Multiple Time Series with One-Sided Dynamic Principal Components.” Unpublished Manuscript.
}


\keyword{ts}
\keyword{package}