% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainODPC.R
\name{cv.odpc}
\alias{cv.odpc}
\title{Automatic choosing of tuning parameters for odpc via cross-validation}
\usage{
cv.odpc(Z, h, k_list = 1:5, max_num_comp = 5, window_size, ncores_k = 1,
  ncores_w = 1, method, tol = 1e-04, niter_max = 500)
}
\arguments{
\item{Z}{Data matrix. Each column is a different time series.}

\item{h}{Forecast horizon.}

\item{k_list}{List of values of k to choose from.}

\item{max_num_comp}{Maximum possible number of components to compute}

\item{window_size}{The size of the rolling window used to estimate the forecasting error.}

\item{ncores_k}{Number of cores to parallelise over the list of ks.}

\item{ncores_w}{Number of cores to parallelise over the rolling window (nested in the ks).}

\item{method}{A string specifying the algorithm used. Options are 'ALS' or 'mix'. See details below.}

\item{tol}{Relative precision. Default is 1e-4.}

\item{niter_max}{Integer. Maximum number of iterations. Default is 500.}
}
\value{
An object of class odpcs, that is, a list of length equal to the number of computed components, each computed using the optimal value of k.
}
\description{
Computes One-Sided Dynamic Principal Components, choosing the number of components and lags automatically, to minimize an estimate
of the forecasting mean squared error.
}
\seealso{
\code{\link{odpc}}, \code{\link{forecast.odpcs}}
}
