% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseODPC.R
\name{cv.sparse_odpc}
\alias{cv.sparse_odpc}
\title{Automatic Choice of Regularization Parameters for Sparse One-Sided Dynamic Principal Components via Cross-Validation}
\usage{
cv.sparse_odpc(
  Z,
  k_list = 1:3,
  max_num_comp = 1,
  nlambda = 20,
  alpha_en = 1,
  ks,
  tol = 1e-04,
  niter_max = 500,
  eps = 0.001,
  ncores = 1
)
}
\arguments{
\item{Z}{Data matrix. Each column is a different time series.}

\item{k_list}{List of values of k to choose from.}

\item{max_num_comp}{Maximum possible number of components to compute.}

\item{nlambda}{Length of penalty sequence.}

\item{alpha_en}{Between 0 and 1, elastic net}

\item{ks}{Optional, number of lags to use.}

\item{tol}{Relative precision. Default is 1e-4.}

\item{niter_max}{Integer. Maximum number of iterations. Default is 500.}

\item{eps}{Between 0 and 1, used to build penalty sequence}

\item{ncores}{Number of cores to use in parallel computations}
}
\value{
An object of class odpcs, that is, a list of length equal to the number of computed components, each computed using the optimal value of k. 
The i-th entry of this list is an object of class \code{odpc}, that is, a list with entries
\item{f}{Coordinates of the i-th dynamic principal component corresponding to the periods \eqn{k_1 + 1,\dots,T}.}
\item{mse}{Mean squared error of the reconstruction using the first i components.}
\item{k1}{Number of lags used to define the i-th dynamic principal component f.}
\item{k2}{Number of lags of f used to reconstruct.}
\item{alpha}{Vector of intercepts corresponding to f.}
\item{a}{Vector that defines the i-th dynamic principal component}
\item{B}{Matrix of loadings corresponding to f. Row number \eqn{k} is the vector of \eqn{k-1} lag loadings.}
\item{call}{The matched call.}
\item{conv}{Logical. Did the iterations converge?}
\item{lambda}{Regularization parameter used for this componen}
\code{components}, \code{fitted}, \code{plot} and \code{print} methods are available for this class.
}
\description{
Computes Sparse One-Sided Dynamic Principal Components, choosing the number of components and regularization parameters automatically, to minimize an estimate
of the forecasting mean squared error.
}
\details{
TBA
}
\examples{
T <- 50 #length of series
m <- 10 #number of series
set.seed(1234)
f <- rnorm(T + 1)
x <- matrix(0, T, m)
u <- matrix(rnorm(T * m), T, m)
for (i in 1:m) {
  x[, i] <- 10 * sin(2 * pi * (i/m)) * f[1:T] + 10 * cos(2 * pi * (i/m)) * f[2:(T + 1)] + u[, i]
}
# Choose parameters to perform a one step ahead forecast. Use 1 or 2 lags, only one component 
# and a window size of 2 (artificially small to keep computation time low). Use two cores for the
# loop over k, two cores for the loop over the window
fit <- cv.sparse_odpc(x, k_list = 1, ncores = 1)
}
\references{
Peña D., Smucler E. and Yohai V.J. (2017). “Forecasting Multiple Time Series with One-Sided Dynamic Principal Components.” Available at https://arxiv.org/abs/1708.04705.
}
\seealso{
\code{\link{odpc}}, \code{\link{crit.odpc}}, \code{\link{forecast.odpcs}}
}
